refactor(selection): migrate caret-related functionality to SelectionController
## Summary
- Updated imports across multiple files to replace references from `CaretModel` to `SelectionController`, enhancing the selection management system.
- Refactored `BlockController`, `EditController`, `KeyboardController`, and `OverlayController` to utilize the new selection structure.
- Introduced new `caretDom` utilities and tests to support caret positioning and selection logic.
- Added comprehensive documentation for the new selection feature and its components.

## Test Plan
- [x] All core tests pass
- [ ] Verify selection behavior and caret positioning in various scenarios
