+++ packages/core/src/features/block/BlockController.ts
@@ -1,9 +1,9 @@
 import type {Range} from '../../shared/editorContracts'
 import {computed, event, watch} from '../../shared/signals'
 import type {DragAction} from '../../shared/types'
-import type {SelectionController} from '../caret/SelectionController'
 import type {Token} from '../parsing'
 import type {TokenModel} from '../parsing/TokenModel'
+import type {SelectionController} from '../selection/SelectionController'
 import type {PropsModel} from '../state/PropsModel'
 import type {ValueModel} from '../state/ValueModel'
 import {createRowContent} from './createRowContent'
