+++ packages/core/src/features/overlay/OverlayController.ts
@@ -3,13 +3,13 @@ import {escape} from '../../shared/escape'
 import {signal, computed, event, effectScope, effect, watch, listen} from '../../shared/signals/index.js'
 import type {Computed} from '../../shared/signals/index.js'
 import type {CoreOption, OverlayMatch, OverlayTrigger, Slot} from '../../shared/types'
-import * as caretDom from '../caret/caretDom'
-import type {SelectionController} from '../caret/SelectionController'
 import type {DomModel} from '../dom/DomModel'
 import type {EditController} from '../edit'
 import type {Token} from '../parsing'
 import {annotate} from '../parsing'
 import type {TokenModel} from '../parsing/TokenModel'
+import * as caretDom from '../selection/caretDom'
+import type {SelectionController} from '../selection/SelectionController'
 import {resolveOverlaySlot} from '../slots'
 import type {OverlaySlot} from '../slots'
 import type {Lifecycle} from '../state/Lifecycle'
