+++ packages/core/src/features/selection/caretDom.spec.ts
