+++ packages/core/src/features/selection/caretDom.ts
