+++ packages/core/src/store/Store.ts
@@ -1,11 +1,11 @@
 import {BlockController, BlockRegistry} from '../features/block'
-import {SelectionController} from '../features/caret'
 import {ClipboardController} from '../features/clipboard'
 import {DomModel} from '../features/dom'
 import {EditController} from '../features/edit'
 import {KeyboardController} from '../features/keyboard'
 import {OverlayController} from '../features/overlay'
 import {MarkFeature, TokenModel} from '../features/parsing'
+import {SelectionController} from '../features/selection'
 import {SlotsFeature} from '../features/slots'
 import {Lifecycle, PropsModel, ValueModel} from '../features/state'
 import {KeyGenerator} from '../shared/classes'
